// mobile/src/api/generated/shop/schemas/wishlistItem.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * ShopDjango API
 * API for ShopDjango project
 * OpenAPI spec version: 1.0.0
 */
import type { ProductList } from './productList';

/**
 * Serializer for WishlistItem model.
 */
export interface WishlistItem {
  readonly id: number;
  readonly product: ProductList;
  product_id: number;
  /** Timestamp when the record was created */
  readonly created_at: string;
}